requirements.txt
streamlink==1.7.0
twitch-python==0.0.19
ipython==7.19.0
requests==2.25.0